+++ lib/src/adapters/_ws_connect_stub.dart
@@ -1,0 +1,5 @@
+import 'package:web_socket_channel/web_socket_channel.dart';
+import '../websocket_config.dart';
+
+WebSocketChannel connectChannel(Uri uri, WebSocketConfig config) =>
+    WebSocketChannel.connect(uri, protocols: config.protocols);