+++ lib/src/compression_io.dart
@@ -1,0 +1,5 @@
+import 'dart:io';
+
+List<int> platformGzipEncode(List<int> data) => gzip.encode(data);
+List<int> platformGzipDecode(List<int> data) => gzip.decode(data);
+const bool platformSupportsCompression = true;