+++ lib/src/compression_stub.dart
@@ -1,0 +1,3 @@
+List<int> platformGzipEncode(List<int> data) => data;
+List<int> platformGzipDecode(List<int> data) => data;
+const bool platformSupportsCompression = false;