+++ lib/websocket_plugin.dart
@@ -1,5 +1,3 @@
-library adapter_websocket;
-
 export 'src/websocket_adapter.dart';
 export 'src/websocket_client.dart';
 export 'src/websocket_config.dart';